+++ tests/callback/demo.c
@@ -13,7 +13,7 @@
  *  0 to stop traversing
  * !0 to continue
  */
-static int sum_numbers_until_null(const json *node, int depth, void *data)
+static int sum_numbers_until_null(const json *node, size_t depth, void *data)
 {
     (void)depth;
     if (json_is_number(node))
